+++ src/Domain/Attributes/Models/Attribute.php
@@ -2,11 +2,11 @@
 
 namespace Dystcz\LunarApi\Domain\Attributes\Models;
 
-use Dystcz\LunarApi\Base\Contracts\Translatable;
-use Dystcz\LunarApi\Hashids\Traits\HashesRouteKey;
+use Dystcz\LunarApi\Domain\Attributes\Concerns\InteractsWithLunarApi;
+use Dystcz\LunarApi\Domain\Attributes\Contracts\Attribute as AttributeContract;
 use Lunar\Models\Attribute as LunarAttribute;
 
-class Attribute extends LunarAttribute implements Translatable
+class Attribute extends LunarAttribute implements AttributeContract
 {
-    use HashesRouteKey;
+    use InteractsWithLunarApi;
 }