+++ src/Domain/Countries/Contracts/Country.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Dystcz\LunarApi\Domain\Countries\Contracts;
+
+use Lunar\Models\Contracts\Country as LunarCountry;
+
+interface Country extends LunarCountry {}