+++ src/Domain/Currencies/Contracts/Currency.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Dystcz\LunarApi\Domain\Currencies\Contracts;
+
+use Lunar\Models\Contracts\Currency as LunarCurrency;
+
+interface Currency extends LunarCurrency {}