+++ src/Domain/Tags/Contracts/Tag.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Dystcz\LunarApi\Domain\Tags\Contracts;
+
+use Lunar\Models\Contracts\Tag as LunarTag;
+
+interface Tag extends LunarTag {}