+++ src/Domain/Transactions/Contracts/Transaction.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Dystcz\LunarApi\Domain\Transactions\Contracts;
+
+use Lunar\Models\Contracts\Transaction as LunarTransaction;
+
+interface Transaction extends LunarTransaction {}