+++ src/Domain/Urls/Contracts/Url.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Dystcz\LunarApi\Domain\Urls\Contracts;
+
+use Lunar\Models\Contracts\Url as LunarUrl;
+
+interface Url extends LunarUrl {}