+++ src/Domain/Users/Concerns/InteractsWithLunarApi.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Dystcz\LunarApi\Domain\Users\Concerns;
+
+use Dystcz\LunarApi\Hashids\Traits\HashesRouteKey;
+
+trait InteractsWithLunarApi
+{
+    use HashesRouteKey;
+}