+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changes
 
+-   Model logic extracted to traits
+-   Added contracts for all models
 -   Added `images` relationship route for `collections`
 -   Added countable relationship tests